Function rclrs::default_initial_value_discriminator

pub fn default_initial_value_discriminator<T: ParameterVariant>(
    available: AvailableValues<'_, T>
) -> Option<T>

The default discriminator that chooses the initial value for a parameter. The implementation here uses a simple preference of

default_value < override_value < prior_value

in ascending order of preference.

The prior_value will automatically be discarded if it is outside the designated range. The override value will not be discarded if it is out of range because that is more likely to be an error that needs to be escalated. You can replace all of this with custom behavior by providing your own discriminator function to ParameterBuilder::discriminate().