Struct rclrs::Node

pub struct Node { /* private fields */ }

A processing unit that can communicate with other nodes.

Nodes are a core concept in ROS 2. Refer to the official “Understanding ROS 2 nodes” tutorial for an introduction.

Ownership of the node is shared with all Publishers and Subscriptions created from it. That means that even after the node itself is dropped, it will continue to exist and be displayed by e.g. ros2 topic as long as its publishers and subscriptions are not dropped.

Naming

A node has a name and a namespace. The node namespace will be prefixed to the node name to form the fully qualified node name. This is the name that is shown e.g. in ros2 node list. Similarly, the node namespace will be prefixed to all names of topics and services created from this node.

By convention, a node name with a leading underscore marks the node as hidden.

It’s a good idea for node names in the same executable to be unique.

Remapping

The namespace and name given when creating the node can be overriden through the command line. In that sense, the parameters to the node creation functions are only the default namespace and name. See also the official tutorial on the command line arguments for ROS nodes, and the Node::namespace() and Node::name() functions for examples.

Rules for valid names

The rules for valid node names and node namespaces are explained in NodeBuilder::new() and NodeBuilder::namespace().

Implementations

impl Node

pub fn get_publisher_names_and_types_by_node( &self, node: &str, namespace: &str ) -> Result<TopicNamesAndTypes, RclrsError>

Returns a list of topic names and types for publishers associated with a node.

pub fn get_subscription_names_and_types_by_node( &self, node: &str, namespace: &str ) -> Result<TopicNamesAndTypes, RclrsError>

Returns a list of topic names and types for subscriptions associated with a node.

pub fn get_service_names_and_types_by_node( &self, node: &str, namespace: &str ) -> Result<TopicNamesAndTypes, RclrsError>

Returns a list of topic names and types for services associated with a node.

pub fn get_client_names_and_types_by_node( &self, node: &str, namespace: &str ) -> Result<TopicNamesAndTypes, RclrsError>

Returns a list of topic names and types for clients associated with a node.

pub fn get_topic_names_and_types( &self ) -> Result<TopicNamesAndTypes, RclrsError>

Returns a list of all topic names and their types.

pub fn get_service_names_and_types( &self ) -> Result<TopicNamesAndTypes, RclrsError>

Returns a list of service names and types for this node.

pub fn get_node_names(&self) -> Result<Vec<NodeNameInfo>, RclrsError>

Returns a list of all node names.

pub fn get_node_names_with_enclaves( &self ) -> Result<Vec<(NodeNameInfo, String)>, RclrsError>

Returns a list of all node names with enclaves.

pub fn count_publishers(&self, topic: &str) -> Result<usize, RclrsError>

Counts the number of publishers for a given topic.

pub fn count_subscriptions(&self, topic: &str) -> Result<usize, RclrsError>

Counts the number of subscriptions for a given topic.

pub fn get_publishers_info_by_topic( &self, topic: &str ) -> Result<Vec<TopicEndpointInfo>, RclrsError>

Returns topic publisher info.

pub fn get_subscriptions_info_by_topic( &self, topic: &str ) -> Result<Vec<TopicEndpointInfo>, RclrsError>

Returns topic subscriptions info.

impl Node

pub fn new(context: &Context, node_name: &str) -> Result<Arc<Node>, RclrsError>

Creates a new node in the empty namespace.

See NodeBuilder::new() for documentation.

pub fn get_clock(&self) -> Clock

Returns the clock associated with this node.

pub fn name(&self) -> String

Returns the name of the node.

This returns the name after remapping, so it is not necessarily the same as the name that was used when creating the node.

Example

// Without remapping
let context = Context::new([])?;
let node = rclrs::create_node(&context, "my_node")?;
assert_eq!(node.name(), "my_node");
// With remapping
let remapping = ["--ros-args", "-r", "__node:=your_node"].map(String::from);
let context_r = Context::new(remapping)?;
let node_r = rclrs::create_node(&context_r, "my_node")?;
assert_eq!(node_r.name(), "your_node");

pub fn namespace(&self) -> String

Returns the namespace of the node.

This returns the namespace after remapping, so it is not necessarily the same as the namespace that was used when creating the node.

Example

// Without remapping
let context = Context::new([])?;
let node =
  rclrs::create_node_builder(&context, "my_node")
  .namespace("/my/namespace")
  .build()?;
assert_eq!(node.namespace(), "/my/namespace");
// With remapping
let remapping = ["--ros-args", "-r", "__ns:=/your_namespace"].map(String::from);
let context_r = Context::new(remapping)?;
let node_r = rclrs::create_node(&context_r, "my_node")?;
assert_eq!(node_r.namespace(), "/your_namespace");

pub fn fully_qualified_name(&self) -> String

Returns the fully qualified name of the node.

The fully qualified name of the node is the node namespace combined with the node name. It is subject to the remappings shown in Node::name() and Node::namespace().

Example

let context = Context::new([])?;
let node =
  rclrs::create_node_builder(&context, "my_node")
  .namespace("/my/namespace")
  .build()?;
assert_eq!(node.fully_qualified_name(), "/my/namespace/my_node");

pub fn create_client<T>( &self, topic: &str ) -> Result<Arc<Client<T>>, RclrsError>
where T: Service,

Creates a Client.

pub fn create_guard_condition(&self) -> Arc<GuardCondition>

Creates a GuardCondition with no callback.

A weak pointer to the GuardCondition is stored within this node. When this node is added to a wait set (e.g. when calling spin_once2 with this node as an argument), the guard condition can be used to interrupt the wait.

pub fn create_guard_condition_with_callback<F>( &mut self, callback: F ) -> Arc<GuardCondition>
where F: Fn() + Send + Sync + 'static,

Creates a GuardCondition with a callback.

A weak pointer to the GuardCondition is stored within this node. When this node is added to a wait set (e.g. when calling spin_once2 with this node as an argument), the guard condition can be used to interrupt the wait.

pub fn create_publisher<T>( &self, topic: &str, qos: QoSProfile ) -> Result<Arc<Publisher<T>>, RclrsError>
where T: Message,

Creates a Publisher.

pub fn create_service<T, F>( &self, topic: &str, callback: F ) -> Result<Arc<Service<T>>, RclrsError>
where T: Service, F: Fn(&rmw_request_id_t, T::Request) -> T::Response + 'static + Send,

Creates a Service.

pub fn create_subscription<T, Args>( &self, topic: &str, qos: QoSProfile, callback: impl SubscriptionCallback<T, Args> ) -> Result<Arc<Subscription<T>>, RclrsError>
where T: Message,

Creates a Subscription.

pub fn domain_id(&self) -> usize

Returns the ROS domain ID that the node is using.

The domain ID controls which nodes can send messages to each other, see the ROS 2 concept article. It can be set through the ROS_DOMAIN_ID environment variable.

Example

// Set default ROS domain ID to 10 here
std::env::set_var("ROS_DOMAIN_ID", "10");
let context = Context::new([])?;
let node = rclrs::create_node(&context, "domain_id_node")?;
let domain_id = node.domain_id();
assert_eq!(domain_id, 10);

pub fn declare_parameter<'a, T: ParameterVariant + 'a>( &'a self, name: impl Into<Arc<str>> ) -> ParameterBuilder<'a, T>

Creates a ParameterBuilder that can be used to set parameter declaration options and declare a parameter as OptionalParameter, MandatoryParameter, or ReadOnly.

Example

let context = Context::new([])?;
let node = rclrs::create_node(&context, "domain_id_node")?;
// Set it to a range of 0-100, with a step of 2
let range = ParameterRange {
    lower: Some(0),
    upper: Some(100),
    step: Some(2),
};
let param = node.declare_parameter("int_param")
                .default(10)
                .range(range)
                .mandatory()
                .unwrap();
assert_eq!(param.get(), 10);
param.set(50).unwrap();
assert_eq!(param.get(), 50);
// Out of range, will return an error
assert!(param.set(200).is_err());

pub fn use_undeclared_parameters(&self) -> Parameters<'_>

Enables usage of undeclared parameters for this node.

Returns a Parameters struct that can be used to get and set all parameters.

pub fn builder(context: &Context, node_name: &str) -> NodeBuilder

Creates a NodeBuilder with the given name.

Convenience function equivalent to NodeBuilder::new().

Example

let context = Context::new([])?;
let node = Node::builder(&context, "my_node").build()?;
assert_eq!(node.name(), "my_node");