Struct rclrs::ParameterBuilder

pub struct ParameterBuilder<'a, T: ParameterVariant> { /* private fields */ }

Builder used to declare a parameter. Obtain this by calling crate::Node::declare_parameter.

Implementations

impl<'a, T: ParameterVariant> ParameterBuilder<'a, T>

pub fn default(self, value: T) -> Self

Sets the default value for the parameter. The parameter value will be initialized to this if no command line override was given for this parameter and if the parameter also had no value prior to being declared.

To customize how the initial value of the parameter is chosen, you can provide a custom function with the method Self::discriminate(). By default, the initial value will be chosen as default_value < override_value < prior_value in order of increasing preference.

pub fn ignore_override(self) -> Self

Ignore any override that was given for this parameter.

If you also use Self::discriminate(), the AvailableValues::override_value field given to the discriminator will be None even if the user had provided an override.

pub fn discard_mismatching_prior_value(self) -> Self

If the parameter was set to a value before being declared with a type that does not match this declaration, discard the prior value instead of emitting a DeclarationError::PriorValueTypeMismatch.

If the type of the prior value does match the declaration, it will still be provided to the discriminator.

pub fn discriminate<F>(self, f: F) -> Self
where F: FnOnce(AvailableValues<'_, T>) -> Option<T> + 'a,

Decide what the initial value for the parameter will be based on the available default_value, override_value, or prior_value.

The default discriminator is default_initial_value_discriminator().

pub fn range(self, range: T::Range) -> Self

Sets the range for the parameter.

pub fn description(self, description: impl Into<Arc<str>>) -> Self

Sets the parameter’s human readable description.

pub fn constraints(self, constraints: impl Into<Arc<str>>) -> Self

Sets the parameter’s human readable constraints. These are not enforced by the library but are displayed on parameter description requests and can be used by integrators to understand complex constraints.

pub fn mandatory(self) -> Result<MandatoryParameter<T>, DeclarationError>

Declares the parameter as a Mandatory parameter, that must always have a value.

See also

Self::optional()

Self::read_only()

pub fn read_only(self) -> Result<ReadOnlyParameter<T>, DeclarationError>

Declares the parameter as a ReadOnly parameter, that cannot be edited.

See also

Self::optional()

Self::mandatory()

pub fn optional(self) -> Result<OptionalParameter<T>, DeclarationError>

Declares the parameter as an Optional parameter, that can be unset.

This will never return the DeclarationError::NoValueAvailable variant.

See also

Self::mandatory()

Self::read_only()

impl<'a, T> ParameterBuilder<'a, Arc<[T]>>
where Arc<[T]>: ParameterVariant,

pub fn default_from_iter( self, default_value: impl IntoIterator<Item = T> ) -> Self

Sets the default for an array-like parameter from an iterable.

impl<'a> ParameterBuilder<'a, Arc<[Arc<str>]>>

pub fn default_string_array<U>(self, default_value: U) -> Self
where U: IntoIterator, U::Item: Into<Arc<str>>,

Sets the default for the parameter from a string-like array.