Trait rclrs::ParameterVariant

pub trait ParameterVariant: Into<ParameterValue> + Clone + TryFrom<ParameterValue> {
    type Range: Into<ParameterRanges> + Default + Clone;

    // Required method
    fn kind() -> ParameterKind;
}

A trait that describes a value that can be converted into a parameter.

Required Associated Types

type Range: Into<ParameterRanges> + Default + Clone

The type used to describe the range of this parameter.

Required Methods

fn kind() -> ParameterKind

Returns the ParameterKind of the implemented type.

Object Safety

This trait is not object safe.